+++ src/utils/__tests__/snapshot-processing.test.ts
@@ -1,25 +1,7 @@
 import { test } from 'vitest';
 import assert from 'node:assert/strict';
-import { buildSnapshotNodeMap } from '@agent-device/contracts/snapshot';
-import { findNearestAncestor, findSnapshotAncestor } from '../../snapshot/snapshot-processing.ts';
+import { findNearestAncestor } from '../../snapshot/snapshot-processing.ts';
 import type { SnapshotNode } from '@agent-device/kernel/snapshot';
-
-test('findSnapshotAncestor walks non-contiguous parent indexes until resolver returns a value', () => {
-  const nodes: SnapshotNode[] = [
-    { ref: 'e10', index: 10, type: 'Window' },
-    { ref: 'e30', index: 30, parentIndex: 20, type: 'Text' },
-    { ref: 'e20', index: 20, parentIndex: 10, type: 'Cell' },
-  ];
-  const visited: number[] = [];
-
-  const ancestor = findSnapshotAncestor(nodes, nodes[1]!, buildSnapshotNodeMap(nodes), (node) => {
-    visited.push(node.index);
-    return node.type === 'Window' ? node : null;
-  });
-
-  assert.deepEqual(visited, [20, 10]);
-  assert.equal(ancestor?.index, 10);
-});
 
 test('findNearestAncestor resolves parents by snapshot index rather than array position', () => {
   const nodes: SnapshotNode[] = [
